requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0

psycopg2-binary==2.9.9  # PostgreSQL adapter

easyocr==1.7.2
pillow==11.2.1
numpy==2.3.0
opencv-python-headless==4.11.0.86
PyMuPDF==1.26.1
regex==2024.11.6

google-generativeai==0.8.5
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
httplib2==0.22.0

requests==2.32.3
